made a seperate usercard component and implemented in facility user page
